+++ dotnet/sample/AutoGen.OpenAI.Sample/Tool_Call_With_Ollama_And_LiteLLM.cs
@@ -8,6 +8,7 @@ using Azure.Core.Pipeline;
 
 namespace AutoGen.OpenAI.Sample;
 
+#region Function
 public partial class Function
 {
     [Function]
@@ -16,6 +17,8 @@ public partial class Function
         return await Task.FromResult("The weather in " + city + " is 72 degrees and sunny.");
     }
 }
+#endregion Function
+
 public class Tool_Call_With_Ollama_And_LiteLLM
 {
     public static async Task RunAsync()
@@ -29,14 +32,7 @@ public class Tool_Call_With_Ollama_And_LiteLLM
         //  - Start LiteLLM with the following command:
         //    - litellm --model ollama_chat/dolphincoder --port 4000
 
-        #region Create_Agent
-        var liteLLMUrl = "http://localhost:4000";
-        using var httpClient = new HttpClient(new CustomHttpClientHandler(liteLLMUrl));
-        var option = new OpenAIClientOptions(OpenAIClientOptions.ServiceVersion.V2024_04_01_Preview)
-        {
-            Transport = new HttpClientTransport(httpClient),
-        };
-
+        # region Create_tools
         var functions = new Function();
         var functionMiddleware = new FunctionCallMiddleware(
             functions: [functions.GetWeatherAsyncFunctionContract],
@@ -44,6 +40,14 @@ public class Tool_Call_With_Ollama_And_LiteLLM
             {
                 { functions.GetWeatherAsyncFunctionContract.Name!, functions.GetWeatherAsyncWrapper },
             });
+        #endregion Create_tools
+        #region Create_Agent
+        var liteLLMUrl = "http://localhost:4000";
+        using var httpClient = new HttpClient(new CustomHttpClientHandler(liteLLMUrl));
+        var option = new OpenAIClientOptions(OpenAIClientOptions.ServiceVersion.V2024_04_01_Preview)
+        {
+            Transport = new HttpClientTransport(httpClient),
+        };
 
         // api-key is not required for local server
         // so you can use any string here
